Isolate dask_array expression integration
